+++ src/__cookbook.rs
@@ -263,7 +263,7 @@
 //! configuration file by the `frame_rate` key.
 //!
 //! #### `clap`
-//! Clap-crate attributes. Available nested attributes: `help`, `long_help`, `short`, `long`,
+//! Clap-crate attributes. Available nested attributes: `help`, `long_help`, `help_heading`, `short`, `long`, `flag`
 //! `flatten`, `subcommand`.
 //! **Note:** the default `long` and `short` values (`#[clap(long)]` and `#[clap(short)]`) is the field name and it's first letter respectively. \
 //! `#[source(clap)]` is equivalent to `#[source(clap(long))]` \